try maizzle to build email html
