+++ .github/workflows/check-dist.yml
@@ -41,11 +41,3 @@ jobs:
             git diff
             exit 1
           fi
-        id: diff
-
-      # If index.js was different than expected, upload the expected version as an artifact
-      - uses: actions/upload-artifact@v2
-        if: ${{ failure() && steps.diff.conclusion == 'failure' }}
-        with:
-          name: dist
-          path: dist/
